refactor: remove unnecessary classes
