+++ backend/app/src/main/java/backend/Router.java
@@ -5,23 +5,12 @@ import static io.javalin.apibuilder.ApiBuilder.post;
 import static io.javalin.apibuilder.ApiBuilder.path;
 import static io.javalin.apibuilder.ApiBuilder.ws;
 
-import backend.controller.GameController;
-import backend.controller.PlayerController;
-import backend.repository.Repository;
-
 // Router handles routing and setting new endpoints
 public class Router {
     private final static Repository repository = new Repository();
 
-    private static final void playerRoutes() {
-        final PlayerController playerController = new PlayerController(repository);
-
-        // Get /players/:id
-        get("{player_id}", ctx -> playerController.getById(ctx, ctx.pathParam("player_id")));
-    }
-
     private static final void gameRoutes() {
-        final GameController gameController = new GameController(repository);
+        final Controller gameController = new Controller(repository);
 
         path("{game_id}", () -> {
             // GET /games/:id
@@ -62,9 +51,6 @@ public class Router {
     }
 
     public static final void getRoutes() {
-        // Set player endpoints
-        path("players", Router::playerRoutes);
-
         // Set game endpoints
         path("games", Router::gameRoutes);
     }
