--- backend/app/src/main/java/backend/repository/Repository.java
@@ -1,12 +1,0 @@
-package backend.repository;
-
-// Repository contains in-memory store
-public class Repository {
-    public final GameRepository games;
-    public final PlayerRepository players;
-
-    public Repository() {
-        this.games = new GameRepository();
-        this.players = new PlayerRepository();
-    }
-}
